// PortalAutoservicio/src/app/models/detalle-comisiones-edd.ts
export class DetalleComisionesedd {
    CLAVESERVIDORPUBLICO: String;
    NOMBRESERVIDORPUBLICO: String;
    APELLIDOPATERNOSERVIDORPUBLICO: String;
    APELLIDOMATERNOSERVIDORPUBLICO: String;
    RFC: String;
    TIPOEVALUACION: String;
    ESTATUSEVALUACION:string;
    IDPROCESOACTUAL:string;
    CORREOSERVIDORPUBLICO:string;
    EVALUACIONKP:string;
}