// PortalAutoservicio/src/app/models/proceso-vigente.ts
export class PorcesoVigente {
    CLAVEPROCESOEVALUACION: string;
    NOMBREPROCESOEVALUACION: string;
    FECHAINICIOPROCESOGENERAL: string;
    FECHAFINPROCESOGENERAL: string;
    FECHAINICIOCAPTURADEMERITOS: string;
    FECHAFINCAPTURADEMERITOS: string;
    DESCRIPCIONPROCESOVIGENTE:string;
    IDPROCESOVIGENTE: string;
}